requirements.txt
Django==2.2.8
Pillow==6.2.0
django-ckeditor==5.8.0